+++ ecommerce/database/migrations/2020_09_10_064710_create_discounts_table.php
@@ -20,7 +20,7 @@ class CreateDiscountsTable extends Migration
             $table->string('type');
             $table->timestamps();
 
-            $table->Integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
         });
     }
