+++ ecommerce/database/migrations/2020_09_11_171753_create_images_table.php
@@ -24,7 +24,7 @@ class CreateImagesTable extends Migration
             $table->string('color');
             $table->timestamps();
 
-            $table->Integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
         });
     }
